reorganize nbconvert
