+++ jupyter_server/serverapp.py
@@ -282,7 +282,7 @@ class ServerWebApplication(web.Application):
         handlers.extend([(r"/logout", settings['logout_handler_class'])])
         handlers.extend(load_handlers('jupyter_server.files.handlers'))
         handlers.extend(load_handlers('jupyter_server.view.handlers'))
-        handlers.extend(load_handlers('jupyter_server.nbconvert.handlers'))
+        handlers.extend(load_handlers('jupyter_server.services.nbconvert.handlers'))
         handlers.extend(load_handlers('jupyter_server.bundler.handlers'))
         handlers.extend(load_handlers('jupyter_server.services.kernelspecs.handlers'))
         handlers.extend(load_handlers('jupyter_server.edit.handlers'))
@@ -291,7 +291,7 @@ class ServerWebApplication(web.Application):
         handlers.extend(load_handlers('jupyter_server.services.kernels.handlers'))
         handlers.extend(load_handlers('jupyter_server.services.contents.handlers'))
         handlers.extend(load_handlers('jupyter_server.services.sessions.handlers'))
-        handlers.extend(load_handlers('jupyter_server.services.nbconvert.handlers'))
+        handlers.extend(load_handlers('jupyter_server.services.nbconvert.api_handlers'))
         handlers.extend(load_handlers('jupyter_server.services.kernelspecs.api_handlers'))
         handlers.extend(load_handlers('jupyter_server.services.security.handlers'))
         handlers.extend(load_handlers('jupyter_server.services.shutdown'))
